Don't forward proto if we receive the header
If we receive a header for X-Forwarded-For and we have the parameter forwardIP, we add the IP of this machine to the list of IPs. However for X-Forwarded-Proto it is not a list. We send the proto of the existing request (if forwardIP is enabled).

This change honors the header if we receive it