Ensure maximum connections per route matches maximum total connections
For a reverse proxy scenario it must be assumed, that many connections
follow the same route. Apache http client limits the connections per
client and per route. This patch ensures, that the maximum number of
connections per route is set to the same value as the maximum total
number of connections.