Improved connection handling (#181)
* Ensure maximum connections per route matches maximum total connections

For a reverse proxy scenario it must be assumed, that many connections
follow the same route. Apache http client limits the connections per
client and per route. This patch ensures, that the maximum number of
connections per route is set to the same value as the maximum total
number of connections.

* Ensure proxy connection is closed when client connection is closed

A chunked response returned by http client is read to completion when
the resulting input stream is closed. For a server send event connection
this can result in an infitite stream.

Closing the http response shutsdown the backing data stream and the
closing can cleanly continue.