Compressed, chunk encoded streams are not flushed immediately
It was observed, that chunk encoded streams were not correctly passed
through the proxy servlet. Analysis showed, that the stream was gzip
compressed.

The issue is located in the Apache http client/JDK: if the stream from
client is compressed, apache http client will delegate to
GzipInputStream. The #available implementation of InflaterInputStream
(parent of GzipInputStream) returns 1 until EOF is reached.

This is not consistent with InputStream#available, which defines:

    A single read or skip of this many bytes will not block, but may
    read or skip fewer bytes.

To work around this, two options are established:

- by default the stream from the backend will not be decompressed and
  passed through as is. In this case the problem is worked around
  because no decompression happens.

- an option is established to configure compression handling by apache
  http client. If this compression handling is enabled, the workaround
  is not to rely on InputStream#available, but flush always.