Fixed Secure Cookie issue where the Proxy is accessible via HTTP.
Fixed issue where proxy servlet is accessible via HTTP but the external service is HTTPS, attempting to pass a cookie marked with the Secure attribute.  In this case, if this attribute remains, the cookie will not be passed on to the client so we must strip it from the cookie.